Split literal.h files into files token.h and tokenstream.h (same for .cpp files) Also remove numbers assigned to TokenTypes
