+++ include/yaramod/builder/yara_expression_builder.h
@@ -10,7 +10,7 @@
 #include <string>
 
 #include "yaramod/types/expression.h"
-#include "yaramod/types/literal.h"
+#include "yaramod/types/tokenstream.h"
 #include "yaramod/types/symbol.h"
 #include "yaramod/yaramod_error.h"
